refactor(common-types): extract @tzurot/config-resolver (PR-2p-2)
Move the config-resolution RESOLVER classes out of the common-types
grab-bag into a dedicated @tzurot/config-resolver package. Continues the
"Slim common-types" epic after the singleton eviction.

What moved (5 classes + their tests): BaseConfigResolver, LlmConfigResolver,
TtsConfigResolver, SttResolver, ConfigCascadeResolver — the cascade
resolution logic. config-resolver depends one-way on common-types (no cycle).

Boundary refinement vs. the plan's "9 files": the config MAPPERS
(Llm/TtsConfigMapper) and TTS provider TYPES (TtsProvider/TtsProviderError)
STAY in common-types — they're data shapes consumed by common-types schemas
(tts-config, generation) and personality loading (PersonalityLoader), so
moving them would cycle. Likewise the config RESULT/CONTRACT types
(ConfigResolutionResult, ResolvedLlmConfig, BaseConfigResolutionResult,
ConfigResolutionSource) moved to common-types/types/configResolution.ts:
bot-client types the gateway /user/llm-config/resolve response against
ConfigResolutionResult and must not depend on the Prisma-backed
config-resolver. Net principle: resolution LOGIC in config-resolver, shared
data SHAPES in common-types.

Consumers repointed: 13 production (6 ai-worker + 7 api-gateway) + ~11 test
files split their imports so the 4 resolver classes come from
@tzurot/config-resolver; cache-invalidation services + mappers + provider
types stay on @tzurot/common-types. Test mocks that stubbed the resolver
classes now mock @tzurot/config-resolver. bot-client + tooling consume zero
resolvers (boundary intact).

Wiring: new package scaffold (mirrors the test-factories template), root +
service tsconfig references, vitest.workspace, knip; depcruise gains a
bot-client→config-resolver ban (Prisma-backed). Files moved via git mv
(history preserved).
