+++ knip.json
@@ -27,6 +27,9 @@
     "packages/clients": {
       "entry": ["src/**/*.test.ts"]
     },
+    "packages/config-resolver": {
+      "entry": ["src/**/*.test.ts"]
+    },
     "packages/tooling": {
       "entry": ["src/**/*.test.ts"],
       "ignoreDependencies": ["eslint"]
