+++ packages/config-resolver/src/LlmConfigResolver.test.ts
@@ -9,18 +9,21 @@
 
 import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
 import { LlmConfigResolver } from './LlmConfigResolver.js';
-import type { LoadedPersonality } from '../types/schemas/index.js';
-import type { PrismaClient } from './prisma.js';
+import type { LoadedPersonality, PrismaClient } from '@tzurot/common-types';
 
 // Mock logger
-vi.mock('../utils/logger.js', () => ({
-  createLogger: () => ({
-    debug: vi.fn(),
-    info: vi.fn(),
-    warn: vi.fn(),
-    error: vi.fn(),
-  }),
-}));
+vi.mock('@tzurot/common-types', async importOriginal => {
+  const actual = await importOriginal<typeof import('@tzurot/common-types')>();
+  return {
+    ...actual,
+    createLogger: () => ({
+      debug: vi.fn(),
+      info: vi.fn(),
+      warn: vi.fn(),
+      error: vi.fn(),
+    }),
+  };
+});
 
 describe('LlmConfigResolver', () => {
   let resolver: LlmConfigResolver;
