+++ services/ai-worker/src/jobs/handlers/pipeline/steps/ConfigStep.test.ts
@@ -13,7 +13,7 @@ import type { Job } from 'bullmq';
 import { JobType, type LLMGenerationJobData, type LoadedPersonality } from '@tzurot/common-types';
 import { ConfigStep } from './ConfigStep.js';
 import type { GenerationContext } from '../types.js';
-import type { ConfigCascadeResolver } from '@tzurot/common-types';
+import type { ConfigCascadeResolver } from '@tzurot/config-resolver';
 
 // Mock common-types logger
 vi.mock('@tzurot/common-types', async importOriginal => {
