+++ services/ai-worker/tsconfig.json
@@ -7,5 +7,8 @@
   },
   "include": ["src/**/*"],
   "exclude": ["node_modules", "dist", "**/*.test.ts", "**/*.spec.ts"],
-  "references": [{ "path": "../../packages/common-types" }]
+  "references": [
+    { "path": "../../packages/common-types" },
+    { "path": "../../packages/config-resolver" }
+  ]
 }
