+++ services/api-gateway/src/routes/user/config-overrides.test.ts
@@ -51,10 +51,6 @@ vi.mock('@tzurot/common-types', async () => {
     getOrCreateUserShell = mockGetOrCreateUserShell;
   }
 
-  class MockConfigCascadeResolver {
-    resolveOverrides = mockResolveOverrides;
-  }
-
   return {
     ...actual,
     createLogger: () => ({
@@ -64,10 +60,18 @@ vi.mock('@tzurot/common-types', async () => {
       error: vi.fn(),
     }),
     UserService: MockUserService,
-    ConfigCascadeResolver: MockConfigCascadeResolver,
     generateUserPersonalityConfigUuid: vi.fn(
       (userId: string, personalityId: string) => `upc-${userId}-${personalityId}`
     ),
+  };
+});
+
+vi.mock('@tzurot/config-resolver', () => {
+  class MockConfigCascadeResolver {
+    resolveOverrides = mockResolveOverrides;
+  }
+  return {
+    ConfigCascadeResolver: MockConfigCascadeResolver,
   };
 });
