+++ services/api-gateway/src/routes/user/personality-config-overrides.ts
@@ -10,7 +10,7 @@
 
 import { Router, type Response, type RequestHandler } from 'express';
 import { StatusCodes } from 'http-status-codes';
-import { ConfigCascadeResolver } from '@tzurot/common-types';
+import { ConfigCascadeResolver } from '@tzurot/config-resolver';
 import { requireUserAuth, requireProvisionedUser } from '../../services/AuthMiddleware.js';
 import { asyncHandler } from '../../utils/asyncHandler.js';
 import {
