+++ services/api-gateway/tsconfig.json
@@ -7,5 +7,9 @@
   },
   "include": ["src/**/*"],
   "exclude": ["node_modules", "dist", "**/*.test.ts", "**/*.spec.ts"],
-  "references": [{ "path": "../../packages/common-types" }, { "path": "../../packages/clients" }]
+  "references": [
+    { "path": "../../packages/common-types" },
+    { "path": "../../packages/config-resolver" },
+    { "path": "../../packages/clients" }
+  ]
 }
