+++ vitest.workspace.ts
@@ -2,6 +2,7 @@ import { defineWorkspace } from 'vitest/config';
 
 export default defineWorkspace([
   'packages/common-types',
+  'packages/config-resolver',
   'packages/clients',
   'packages/test-factories',
   'services/ai-worker',
